Wizard: Rename `Empty` to `EmptyUserState`
This renames `Empty` to `EmptyUserState` to make the name more descriptive.
